Add VoucherButtons widget for voucher screen
